# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  messages: ChatMessage[];
  conversationId?: string | null;
}

export interface ChatResponse {
  message: string;
  role: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Conversation {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type StoredMessageRole = typeof StoredMessageRole[keyof typeof StoredMessageRole];


export const StoredMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface StoredMessage {
  id: number;
  conversationId: string;
  role: StoredMessageRole;
  content: string;
  createdAt: string;
}

export interface UpdateConversationInput {
  title: string;
}

export interface MemoryItem {
  key: string;
  value: string;
  label: string;
  updatedAt: string;
}

export interface UpsertMemoryInput {
  key: string;
  value: string;
  label: string;
}