# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory:
    get:
      operationId: listMemory
      tags: [memory]
      summary: Get all stored user memory facts
      responses:
        "200":
          description: All memory facts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryItem"
    post:
      operationId: upsertMemory
      tags: [memory]
      summary: Create or update a memory fact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertMemoryInput"
      responses:
        "200":
          description: Saved memory fact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"
  /memory/{key}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory fact by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    MemoryItem:
      type: object
      required: [key, value, label, updatedAt]
      properties:
        key:
          type: string
        value:
          type: string
        label:
          type: string
        updatedAt:
          type: string
          format: date-time
    UpsertMemoryInput:
      type: object
      required: [key, value, label]
      properties:
        key:
          type: string
        value:
          type: string
        label:
          type: string