# lib/api-zod/src/generated/types/memoryItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryItem {
  key: string;
  value: string;
  label: string;
  updatedAt: Date;
}